server/requirements.txt
astroid==1.6.1
dj-database-url==0.4.2
Django==2.0.2
django-crispy-forms==1.6.1
django-filter==1.1.0
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
isort==4.3.3
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
PyJWT==1.5.3
pylint==1.8.2
pytz==2017.2
six==1.11.0
wrapt==1.10.11
